requirements.txt
black==22.10.0
click==8.1.3
colorama==0.4.6
cx-Oracle==8.3.0
flake8==5.0.4
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
